Move package private mapper test to MappersTest
Require to move package-private classes to the same package as the test
to make them visible.
