+++ core-common/src/test/java/org/mapstruct/factory/MappersTest.java
@@ -47,4 +47,9 @@ public class MappersTest {
         assertThat( Mappers.getMapper( SomeClass.Foo.class ) ).isNotNull();
         assertThat( Mappers.getMapper( SomeClass.NestedClass.Foo.class ) ).isNotNull();
     }
+
+    @Test
+    public void shouldReturnPackagePrivateImplementationInstance() {
+        assertThat( Mappers.getMapper( PackagePrivateMapper.class ) ).isNotNull();
+    }
 }
