+++ core-common/src/test/java/org/mapstruct/factory/PackagePrivateMapper.java
@@ -16,7 +16,8 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.mapstruct.test;
+package org.mapstruct.factory;
 
-class PackagePrivateMapperImpl implements PackagePrivateMapper {
+interface PackagePrivateMapper {
+
 }
